fix: reporterinfo port, null condition check
