+++ src/main/java/com/innovactions/incident/port/outbound/ChannelAdministrationPort.java
@@ -2,8 +2,6 @@ package com.innovactions.incident.port.outbound;
 
 import java.util.List;
 
-import com.innovactions.incident.adapter.outbound.SlackChannelAdministrationAdapter.ReporterInfo;
-
 /**
  * Port interface for Slack channel administration tasks (Create channel, set topic, invite/kick
  * users).
