+++ src/main/java/com/innovactions/incident/port/outbound/ReporterInfo.java
@@ -1,0 +1,18 @@
+package com.innovactions.incident.port.outbound;
+
+/**
+ * Represents reporter information extracted from a channel topic.
+ *
+ * <p>This is a domain value object used by the ChannelAdministrationPort interface
+ * and its implementations. It contains the reporter's ID and the platform they used.
+ */
+public class ReporterInfo {
+  public final String reporterId;
+  public final String platform;
+
+  public ReporterInfo(String reporterId, String platform) {
+    this.reporterId = reporterId;
+    this.platform = platform;
+  }
+}
+
